requirements.txt
# OmniParse - Python Dependencies
# Generated from pyproject.toml

# Core dependencies
scikit-learn>=1.3.2
Pillow>=10.1.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
transformers==4.41.2
numpy>=1.26.1
python-dotenv>=1.0.0
torch>=2.2.2
tqdm>=4.66.1
tabulate>=0.9.0
ftfy>=6.1.1
texify>=0.1.8
rapidfuzz>=3.8.1
surya-ocr>=0.4.3
filetype>=1.2.0
regex>=2024.4.28
pdftext>=0.3.10
grpcio>=1.63.0
fastapi>=0.111.0
uvicorn>=0.29.0
pypdfium2>=4.30.0
moviepy==1.0.3
openai-whisper>=20231117
pytube>=15.0.0
beautifulsoup4>=4.12.3
html2text>=2024.2.26
selenium>=4.21.0
webdriver-manager>=4.0.1
img2pdf>=0.5.1
matplotlib>=3.9.0
timm>=1.0.7
art>=6.2
gradio>=4.37.1
nltk>=3.8.1
marker-pdf==0.2.16
einops>=0.7.0
# flash-attn removed - requires CUDA compilation and not compatible with Docker build
